requirements.txt
Django==1.10.2
django-autoslug==1.9.3
gunicorn==19.6.0
mysqlclient==1.3.7
Pillow==3.3.1
pkg-resources==0.0.0
selenium==2.53.6
tblib==1.3.0
